% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opentims.R
\name{MS2}
\alias{MS2}
\title{Get MS2 frame numbers.}
\usage{
MS2(opentims)
}
\arguments{
\item{opentims}{Instance of OpenTIMS}
}
\value{
Numbers of frames corresponding to MS2, i.e. fragment ions.
}
\description{
Get MS2 frame numbers.
}
