% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opentims.R
\docType{class}
\name{OpenTIMS-class}
\alias{OpenTIMS-class}
\title{TimsTOF data accessor.}
\description{
S4 class that facilitates data queries for TimsTOF data.
}
\section{Slots}{

\describe{
\item{\code{analysis.tdf_bin}}{Path to raw data (typically *.d/analysis.tdf_bin).}

\item{\code{analysis.tdf}}{Path to sqlite database (typically *.d/analysis.tdf).}

\item{\code{handle}}{Pointer to raw data.}

\item{\code{min_frame}}{The index of the minimal frame.}

\item{\code{max_frame}}{The index of the miximal frame.}

\item{\code{frames}}{A data.frame with information on the frames (contents of the Frames table in the sqlite db).}

\item{\code{windows}}{A data.frame with information on the windows.}

\item{\code{MS1}}{Indices of frames corresponding to MS1, i.e. precursor ions.}

\item{\code{MS2}}{Indices of frames corresponding to MS2, i.e. fragment ions.}

\item{\code{analysis.tdf.table_names}}{Names of tables in the accompanying sqlite database.}
}}

