% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opentims.R
\name{OpenTIMS}
\alias{OpenTIMS}
\title{Get OpenTIMS data representation.}
\usage{
OpenTIMS(path, analysis.tdf = "")
}
\arguments{
\item{path}{Path to the TimsTOF .d folder. Otherwise direct path to .tdf_bin file containing the raw data.}

\item{db_path}{Path to the sqlite database with the .tdf extension. Only supply it when 'path' corresponded to raw data file. If 'path' corresponds to the whole folder, you can skip this argument.}
}
\description{
Get OpenTIMS data representation.
}
\examples{
\dontrun{
D = OpenTIMS(path_to_.d_folder)
D[1] # First frame.
}
}
