% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opentims.R
\name{query}
\alias{query}
\title{Query for raw data.}
\usage{
query(
  opentims,
  frames = NULL,
  from = NULL,
  to = NULL,
  by = 1L,
  columns = c("frame", "scan", "tof", "intensity", "mz", "dt", "rt")
)
}
\arguments{
\item{opentims}{Instance of OpenTIMS.}

\item{frames}{Vector of frame numbers to extract.}

\item{from}{First frame to select.}

\item{to}{Last frame to select.}

\item{by}{Each 'by'-th frame will be selected. Default to each frame (1).}

\item{columns}{Vector of columns to extract. Defaults to all columns.}
}
\value{
data.frame with selected columns.
}
\description{
Get the raw data from Bruker's 'tdf_bin' format.
Extract either by frame numbers ('frames') or by starting frame ('from'), ending frame (inclusive, 'end'), and step in frames ('by'), like in the 'seq' function.
If 'frames' and one of the end will not be specified, the minimal or maximal frame will be chosen instead.
If both 'frames', 'from' and 'end' are left at default, all of the data will be copied into RAM. Like ALL, so don't do it, unless you have a lot of RAM.
Defaults to both raw data ('frame','scan','tof','intensity') and its tranformations into physical units ('mz','dt','rt').
}
