% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opentims.R
\name{range,OpenTIMS-method}
\alias{range,OpenTIMS-method}
\title{Select a range of frames to extract.}
\usage{
\S4method{range}{OpenTIMS}(x, start, stop, step = 1L)
}
\arguments{
\item{x}{OpenTIMS data instance.}

\item{start}{The first frame to extract.}

\item{stop}{The last+1 frame to extract. Frame with that number will not get extracted, but some below that number might.}

\item{step}{The step}
}
\description{
This is similar to using the start:stop:step operator in Python.
}
